+++ Anarchy/ChangeLog.md
@@ -2,3 +2,6 @@
 * Prevent subelements from being overriden while default tool is active. Important while using EDT transform menu while near subelements.
 * Anarchy components tool now works on subelements. 
 * While Selecting subelements with Debug Toggle you can now add or remove Anarchy or Elevation Lock components.
+* Added very limited and basic game pad support. Just Anarchy toggle.
+* Removed I18N Everywhere dependency. 
+* Translations for officially supported languages are handled internally with embedded resource files.